Extract Connection implementation from DriverRemoteConnection
Extract Connection implementation out of the DriverRemoteConnection.
The DriverRemoteConnection uses a Client instance, a Client instance
may contain one or more Connection instances.
