+++ gremlin-javascript/src/main/javascript/gremlin-javascript/lib/driver/remote-connection.js
@@ -25,23 +25,40 @@
 const t = require('../process/traversal');
 const TraversalStrategy = require('../process/traversal-strategy').TraversalStrategy;
 
+/**
+ * Represents an abstraction of a "connection" to a "server" that is capable of processing a traversal and
+ * returning results.
+ */
 class RemoteConnection {
   constructor(url) {
     this.url = url;
   }
 
   /**
-   * @abstract
-   * @param {Bytecode} bytecode
-   * @param {String} op Operation to perform, defaults to bytecode.
-   * @param {Object} args The arguments for the operation. Defaults to an associative array containing values for "aliases" and "gremlin" keys.
-   * @param {String} requestId A requestId for the current request. If none provided then a requestId is generated internally.
-   * @param {String} processor The processor to use on the connection.
+   * Opens the connection, if its not already opened.
    * @returns {Promise}
    */
-  submit(bytecode, op, args, requestId, processor) {
-    throw new Error('submit() was not implemented');
+  open() {
+    throw new Error('open() must be implemented');
+  }
+
+  /**
+   * Submits the <code>Bytecode</code> provided and returns a <code>RemoteTraversal</code>.
+   * @abstract
+   * @param {Bytecode} bytecode
+   * @returns {Promise} Returns a <code>Promise</code> that resolves to a <code>RemoteTraversal</code>.
+   */
+  submit(bytecode) {
+    throw new Error('submit() must be implemented');
   };
+
+  /**
+   * Closes the connection, if its not already opened.
+   * @returns {Promise}
+   */
+  close() {
+    throw new Error('close() must be implemented');
+  }
 }
 
 class RemoteTraversal extends t.Traversal {
@@ -75,8 +92,4 @@ class RemoteStrategy extends TraversalStrategy {
   }
 }
 
-module.exports = {
-  RemoteConnection: RemoteConnection,
-  RemoteStrategy: RemoteStrategy,
-  RemoteTraversal: RemoteTraversal
-};
+module.exports = { RemoteConnection, RemoteStrategy, RemoteTraversal };
