+++ gremlin-javascript/src/main/javascript/gremlin-javascript/test/helper.js
@@ -21,18 +21,18 @@
  * @author Jorge Bay Gondra
  */
 'use strict';
-const os = require('os');
 
 const DriverRemoteConnection = require('../lib/driver/driver-remote-connection');
 const Client = require('../lib/driver/client');
 const PlainTextSaslAuthenticator = require('../lib/driver/auth/plain-text-sasl-authenticator');
 
+/** @returns {DriverRemoteConnection} */
 exports.getConnection = function getConnection(traversalSource) {
   return new DriverRemoteConnection('ws://localhost:45940/gremlin', { traversalSource: traversalSource });
 };
 
-exports.getSecureConnectionWithPlainTextSaslAuthenticator = function getConnection(traversalSource) {
-  const authenticator = new PlainTextSaslAuthenticator('stephen', 'password');
+exports.getSecureConnectionWithPlainTextSaslAuthenticator = (traversalSource, username, password) => {
+  const authenticator = new PlainTextSaslAuthenticator(username, password);
   return new DriverRemoteConnection('ws://localhost:45941/gremlin', { 
     traversalSource: traversalSource, 
     authenticator: authenticator,
